requirements.txt
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
idna==3.2
inline==0.0.1
keyboard==0.13.5
markup==0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyTelegramBotAPI==3.7.2
pyTelegramBotAPI-markups==0.0.4
PyYAML==5.4.1
reply==0.2
requests==2.26.0
service-identity==21.1.0
six==1.16.0
telebot-utils==0.0.8
TelegramBotAPI==0.3.2
trafaret==2.1.0
trafaret-config==2.0.2
urllib3==1.26.6
